--- a/core/trino-main/src/main/java/io/trino/execution/buffer/PagesSerdeFactory.java
+++ b/core/trino-main/src/main/java/io/trino/execution/buffer/PagesSerdeFactory.java
@@ -13,6 +13,12 @@
  */
 package io.trino.execution.buffer;
 
+import io.airlift.compress.Compressor;
+import io.airlift.compress.Decompressor;
+import io.airlift.compress.lz4.Lz4Compressor;
+import io.airlift.compress.lz4.Lz4Decompressor;
+import io.airlift.compress.zstd.ZstdCompressor;
+import io.airlift.compress.zstd.ZstdDecompressor;
 import io.trino.spi.block.BlockEncodingSerde;
 
 import javax.crypto.SecretKey;
@@ -26,21 +32,39 @@ public class PagesSerdeFactory
     private static final int SERIALIZED_PAGE_DEFAULT_BLOCK_SIZE_IN_BYTES = 64 * 1024;
 
     private final BlockEncodingSerde blockEncodingSerde;
-    private final boolean compressionEnabled;
+    private final CompressionCodec compressionCodec;
 
-    public PagesSerdeFactory(BlockEncodingSerde blockEncodingSerde, boolean compressionEnabled)
+    public PagesSerdeFactory(BlockEncodingSerde blockEncodingSerde, CompressionCodec compressionCodec)
     {
         this.blockEncodingSerde = requireNonNull(blockEncodingSerde, "blockEncodingSerde is null");
-        this.compressionEnabled = compressionEnabled;
+        this.compressionCodec = requireNonNull(compressionCodec, "compressionCodec is null");
     }
 
     public PageSerializer createSerializer(Optional<SecretKey> encryptionKey)
     {
-        return new PageSerializer(blockEncodingSerde, compressionEnabled, encryptionKey, SERIALIZED_PAGE_DEFAULT_BLOCK_SIZE_IN_BYTES);
+        return new PageSerializer(blockEncodingSerde, createCompressor(compressionCodec), encryptionKey, SERIALIZED_PAGE_DEFAULT_BLOCK_SIZE_IN_BYTES);
     }
 
     public PageDeserializer createDeserializer(Optional<SecretKey> encryptionKey)
     {
-        return new PageDeserializer(blockEncodingSerde, compressionEnabled, encryptionKey, SERIALIZED_PAGE_DEFAULT_BLOCK_SIZE_IN_BYTES);
+        return new PageDeserializer(blockEncodingSerde, createDecompressor(compressionCodec), encryptionKey, SERIALIZED_PAGE_DEFAULT_BLOCK_SIZE_IN_BYTES);
+    }
+
+    public static Optional<Compressor> createCompressor(CompressionCodec compressionCodec)
+    {
+        return switch (compressionCodec) {
+            case NONE -> Optional.empty();
+            case LZ4 -> Optional.of(new Lz4Compressor());
+            case ZSTD -> Optional.of(new ZstdCompressor());
+        };
+    }
+
+    public static Optional<Decompressor> createDecompressor(CompressionCodec compressionCodec)
+    {
+        return switch (compressionCodec) {
+            case NONE -> Optional.empty();
+            case LZ4 -> Optional.of(new Lz4Decompressor());
+            case ZSTD -> Optional.of(new ZstdDecompressor());
+        };
     }
 }