… core/trino-main/src/test/java/io/trino/execution/buffer/BenchmarkPagesSerde.java
… core/trino-main/src/test/java/io/trino/execution/buffer/TestPagesSerde.java
… core/trino-main/src/test/java/io/trino/execution/buffer/TestingPagesSerdeFactory.java
… core/trino-main/src/test/java/io/trino/operator/output/BenchmarkPartitionedOutputOperator.java
… core/trino-main/src/test/java/io/trino/operator/output/TestPagePartitioner.java
… core/trino-main/src/test/java/io/trino/operator/spiller/BenchmarkBinaryFileSpiller.java
… core/trino-main/src/test/java/io/trino/spiller/TestBinaryFileSpiller.java
… core/trino-main/src/test/java/io/trino/spiller/TestFileSingleStreamSpiller.java
… core/trino-main/src/test/java/io/trino/spiller/TestFileSingleStreamSpillerFactory.java
… core/trino-main/src/test/java/io/trino/spiller/TestNodeSpillConfig.java
… core/trino-main/src/test/java/io/trino/sql/analyzer/TestFeaturesConfig.java
… docs/src/main/sphinx/admin/properties-exchange.md
… docs/src/main/sphinx/admin/properties-spilling.md