@wendigo 2024-01-04
I'm concerned that zstd is using jni which could affect gc @hackeryang