@wendigo 2024-01-04
I don't think that we should switch by default. We could make it configurable