@wendigo 2024-01-04
@raunaqmorarka I've assumed that zstd codec is using zstd-jni underneath, my mistake