@hackeryang 2024-01-10
@raunaqmorarka @electrum Thank you very much for the benchmarks and advice~ I have separated two commits for exchanges and spills, deprecated `boolean compressionEnabled` to use enum `CompressionCodec.NONE`, please review again.

I didn't figure out whether I should modify the serialize size in the test:  
![image](https://github.com/trinodb/trino/assets/26461591/58e3728a-dc63-448a-9ad7-c5697368eead)