@hackeryang 2024-01-18
> > > > It seems CI has been passed(except the irrelevant errors), @raunaqmorarka can you please help merge, thanks~
> > > 
> > > 
> > > Please fix the compile error on the first commit
> > 
> > 
> > @raunaqmorarka I have fixed the error, please cc again~
> 
> We still need a fix to the failing test TestJdbcConnection.testSession

@raunaqmorarka Thanks for pointing this out~ I have fixed this test failure, please review again~  

The error this time is about Kudu, seems not related to our PR